idk what my commit messages are anymore

+++ multiplayer/game_loop.go
@@ -23,7 +23,7 @@ type ClientPlayer struct {
 }
 
 const (
-	NewGuess       int = iota
+	NewGuess       uint8 = iota
 	NewEffect      
 	PlayerList     
 	PlayerDied
